move MappingRule to mapping.Rule
